+++ test/support/conn_case.ex
@@ -1,0 +1,9 @@
+defmodule Bolt.Sips.ConnCase do
+  use ExUnit.CaseTemplate
+
+  setup_all do
+    conn = Bolt.Sips.conn()
+
+    {:ok, conn: conn}
+  end
+end